+++ CHANGELOG.md
@@ -26,6 +26,9 @@
 - Add `Valkey` field `userConfig.frequent_snapshots`, type `boolean`: When enabled, Valkey will create
   frequent local RDB snapshots
 - Add kind: `AlloyDBOmni`
+- Change `OpenSearch` field `userConfig.opensearch.auth_failure_listeners.internal_authentication_backend_limiting.allowed_tries`:
+  maximum ~~`2147483647`~~ → `32767`
+- Change `OpenSearch` field `userConfig.opensearch.auth_failure_listeners.ip_rate_limiting`: deprecated
 
 ## v0.27.0 - 2025-01-16