+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
 
+## v0.28.0 - 2025-02-17
+
+- Add kind: `AlloyDBOmni`
 - Deprecate `Redis`: use `Valkey` instead. Please follow [these](https://aiven.io/docs/products/caching/howto/upgrade-aiven-for-caching-to-valkey#upgrade-service) instructions to upgrade your service to Valkey
 - Deprecate `Cassandra`, see Aiven platform [end-of-life](https://aiven.io/docs/platform/reference/end-of-life) policy.
 - Change `Cassandra` field `userConfig.ip_filter`: maxItems ~~`1024`~~ → `2048`
@@ -25,7 +28,6 @@
 - Change `Valkey` field `userConfig.ip_filter`: maxItems ~~`1024`~~ → `2048`
 - Add `Valkey` field `userConfig.frequent_snapshots`, type `boolean`: When enabled, Valkey will create
   frequent local RDB snapshots
-- Add kind: `AlloyDBOmni`
 - Change `OpenSearch` field `userConfig.opensearch.auth_failure_listeners.internal_authentication_backend_limiting.allowed_tries`:
   maximum ~~`2147483647`~~ → `32767`
 - Change `OpenSearch` field `userConfig.opensearch.auth_failure_listeners.ip_rate_limiting`: deprecated